Move actor-discovery utils to `.service._registry
